VxDesign: Split out Districts/Precincts screens from Geography screen
